fix: applied game content optimizations

+++ Server.Reawakened/BundleHost/Controllers/AssetHostController.cs
@@ -12,6 +12,8 @@ using Server.Reawakened.BundleHost.Extensions;
 using Server.Reawakened.BundleHost.Models;
 using Server.Reawakened.BundleHost.Services;
 using FileIO = System.IO.File;
+using System.Collections.Concurrent;
+using System.Threading;
 
 namespace Server.Reawakened.BundleHost.Controllers;
 
@@ -19,6 +21,7 @@ namespace Server.Reawakened.BundleHost.Controllers;
 public class AssetHostController(BuildAssetList buildAssetList, ILogger<AssetHostController> logger,
     AssetBundleRConfig config, BuildXmlFiles buildXmlList) : Controller
 {
+    private static readonly ConcurrentDictionary<string, SemaphoreSlim> _bundleLocks = new(StringComparer.OrdinalIgnoreCase);
     [HttpGet]
     public async Task<IActionResult> GetAsset([FromRoute] string folder, [FromRoute] string file)
     {
@@ -73,53 +76,73 @@ public class AssetHostController(BuildAssetList buildAssetList, ILogger<AssetHos
 
         if (!FileIO.Exists(bundlePath) || config.AlwaysRecreateBundle)
         {
-            if (config.LogAssetLoadInfo)
-                logger.LogInformation(
-                    "Creating Bundle {Name} from {Time} [{Type}]",
-                    assetName,
-                    DateTime.UnixEpoch.AddSeconds(asset.CacheTime).ToShortDateString(),
-                    config.AlwaysRecreateBundle ? "FORCED" : "NOT EXIST"
-                );
+            var semaphore = _bundleLocks.GetOrAdd(bundlePath, _ => new SemaphoreSlim(1, 1));
+            await semaphore.WaitAsync();
 
-            using var stream = new MemoryStream();
-            var writer = new EndianWriter(stream, EndianType.BigEndian);
-
-            var unityVersion = new UnityVersion(asset.UnityVersion);
-
-            var fileName = Path.GetFileName(asset.Path);
-
-            var data = new FixedAssetFile();
-            await data.ReadAsync(asset.Path);
-
-            var metadata = new BundleMetadata(fileName, data.FileSize);
-            metadata.FixMetadata((uint)metadata.GetEndianSize());
-
-            var header = new RawBundleHeader(data.FileSize, metadata.MetadataSize, unityVersion);
-            header.FixHeader((uint)header.GetEndianSize());
-
-            header.Write(writer);
-            metadata.Write(writer);
-            data.Write(writer);
-
-            // WRITE
             try
             {
-                await FileIO.WriteAllBytesAsync(bundlePath, stream.ToArray());
+                if (FileIO.Exists(bundlePath) && !config.AlwaysRecreateBundle)
+                    return bundlePath;
+
+                if (config.LogAssetLoadInfo)
+                    logger.LogInformation(
+                        "Creating Bundle {Name} from {Time} [{Type}]",
+                        assetName,
+                        DateTime.UnixEpoch.AddSeconds(asset.CacheTime).ToShortDateString(),
+                        config.AlwaysRecreateBundle ? "FORCED" : "NOT EXIST"
+                    );
+
+                var tempPath = bundlePath + ".tmp";
+
+                await using (var fileStream = new FileStream(tempPath, FileMode.Create, FileAccess.Write, FileShare.None, 1024 * 64, true))
+                {
+                    var writer = new EndianWriter(fileStream, EndianType.BigEndian);
+
+                    var unityVersion = new UnityVersion(asset.UnityVersion);
+
+                    var fileName = Path.GetFileName(asset.Path);
+
+                    var data = new FixedAssetFile();
+                    await data.ReadAsync(asset.Path);
+
+                    var metadata = new BundleMetadata(fileName, data.FileSize);
+                    metadata.FixMetadata((uint)metadata.GetEndianSize());
+
+                    var header = new RawBundleHeader(data.FileSize, metadata.MetadataSize, unityVersion);
+                    header.FixHeader((uint)header.GetEndianSize());
+
+                    header.Write(writer);
+                    metadata.Write(writer);
+                    data.Write(writer);
+
+                    await fileStream.FlushAsync();
+
+                    if (config.DebugInfo)
+                    {
+                        await FileIO.WriteAllTextAsync($"{basePath}.headerVars", JsonConvert.SerializeObject(header, Formatting.Indented));
+                        await FileIO.WriteAllBytesAsync($"{basePath}.header", header.GetEndian());
+
+                        await FileIO.WriteAllTextAsync($"{basePath}.metadataVars",
+                            JsonConvert.SerializeObject(metadata, Formatting.Indented));
+                        await FileIO.WriteAllBytesAsync($"{basePath}.metadata", metadata.GetEndian());
+                        FileIO.Copy(asset.Path!, $"{basePath}.cache", true);
+                    }
+                }
+
+                // Atomic replace
+                if (FileIO.Exists(bundlePath))
+                    FileIO.Delete(bundlePath);
+
+                FileIO.Move(tempPath, bundlePath);
             }
-            catch (IOException)
+            catch (Exception ex)
             {
-
+                logger.LogError(ex, "Error creating bundle for {AssetName} at {BundlePath}", assetName, bundlePath);
+                throw;
             }
-
-            if (config.DebugInfo)
+            finally
             {
-                await FileIO.WriteAllTextAsync($"{basePath}.headerVars", JsonConvert.SerializeObject(header, Formatting.Indented));
-                await FileIO.WriteAllBytesAsync($"{basePath}.header", header.GetEndian());
-
-                await FileIO.WriteAllTextAsync($"{basePath}.metadataVars",
-                    JsonConvert.SerializeObject(metadata, Formatting.Indented));
-                await FileIO.WriteAllBytesAsync($"{basePath}.metadata", metadata.GetEndian());
-                FileIO.Copy(asset.Path!, $"{basePath}.cache", true);
+                semaphore.Release();
             }
         }
